+++ lib/src/migrators/module.dart
@@ -252,11 +252,23 @@ class _ModuleMigrationVisitor extends MigrationVisitor {
           _shouldForward(declaration.name, forImportOnly: true) && !private;
       if (!visibleAtEntrypoint && !shouldBeVisible) continue;
 
-      var url = visibleAtEntrypoint ? entrypoint : declaration.sourceUrl;
-      var prefix = renamedMembers.containsKey(declaration) ||
-              (visibleAtEntrypoint && _startsWithPrefix(declaration.name))
-          ? prefixToRemove
-          : declaration.forward?.prefix ?? '';
+      Uri url;
+      if (visibleAtEntrypoint) {
+        url = entrypoint;
+      } else if (declaration.isImportOnly) {
+        url = declaration.importOnlyUrl;
+      } else {
+        url = declaration.sourceUrl;
+      }
+
+      var prefix = '';
+      if (renamedMembers.containsKey(declaration) ||
+          (visibleAtEntrypoint && _startsWithPrefix(declaration.name))) {
+        prefix = prefixToRemove;
+      } else if (declaration.forward?.prefix != null &&
+          url == declaration.forwardedUrl) {
+        prefix = declaration.forward.prefix;
+      }
       forwardsByUrl
           .putIfAbsent(url, () => {})
           .putIfAbsent(prefix, () => {})